requirements.txt
openai
python-dotenv
python-docx
pillow
matplotlib
tqdm
